Some game tweaks
I changed the display of the results screen to make it less cluttered
and to make it look ok on my phone.

I changed the way that we're using the back stack so that different
buttons make more sense when pressed. For instance, now when you
press back from the game screen, it takes you to the home screen.
The same is now true of the results screen as well. This is better
than quitting out of the app, and then having to reload it.

I set up the state save/load, so that changing the orientation would
preserve the game state. This will also keep the game state when
android closes our app temporarily.

I removed the option for landscape mode via modifications to the
manifest.

I fixed the newGame() method (actually I renamed it). It seems that
we were using two different features and sort of putting them
near each other. One, I've called softReset(), is a visual reset.
The second, I've called hardReset() is a softReset(), and then
the game logic is also reset, and the game activity's own
internal logic is reset.

This allowed me to add a menu on the top right. The menu allows for
a 'New Game' to be started without leaving the activity.